requirements.txt
acme==2.9.0
alembic==1.16.4
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.7.1
asyncpg==0.29.0
attrs==23.1.0
build==1.2.1
CacheControl==0.14.0
certifi==2025.7.14
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.1.0
cleo==2.1.0
click==8.2.1
ConfigArgParse==1.7
configobj==5.0.8
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
coverage==7.4.0
crashtest==0.4.1
cryptography==42.0.5
cssselect2==0.7.0
datadispatch==1.0.0
distlib==0.3.8
distro==1.9.0
django-formtools==2.4.1
django-js-asset==2.1.0
django-mptt==0.14.0
django-polymorphic==3.1.0
django-sekizai==4.1.0
djangocms-admin-style==3.2.4
dnspython==2.7.0
drf-excel==2.3.0
drf-extensions==0.7.1
drf-spectacular==0.26.2
drf-spectacular-sidecar==2023.5.1
dulwich==0.21.7
easy-thumbnails==2.8.5
ecdsa==0.19.1
email_validator==2.2.0
et-xmlfile==1.1.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
fastjsonschema==2.19.1
filelock==3.13.3
flake8==7.0.0
greenlet==3.2.3
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
installer==0.7.0
itsdangerous==2.2.0
itypes==1.2.0
jaraco.classes==3.4.0
Jinja2==3.1.6
josepy==1.14.0
jsonschema==4.17.3
jsonschema-specifications==2023.12.1
keyring==24.3.1
lxml==4.9.3
Mako==1.3.10
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
more-itertools==10.2.0
msgpack==1.0.8
mypy-extensions==1.0.0
openpyscad==0.5.0
openpyxl==3.1.2
orjson==3.11.1
packaging==23.1
parsedatetime==2.6
passlib==1.7.4
pexpect==4.9.0
pillow==10.2.0
pkginfo==1.10.0
platformdirs==4.2.0
pluggy==1.5.0
poetry==1.8.2
poetry-core==1.9.0
poetry-plugin-export==1.7.1
# psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pyasn1==0.6.1
pycodestyle==2.11.1
pycparser==2.21
pydantic==2.11.7
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyflakes==3.2.0
Pygments==2.19.2
pyOpenSSL==24.0.0
pyparsing==3.1.1
pyproject_hooks==1.0.0
pyRFC3339==1.1
pyrsistent==0.19.3
pytest==8.3.2
pytest-asyncio==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.2
RapidFuzz==3.10.0
referencing==0.32.0
reportlab==4.0.4
requests==2.31.0
requests-toolbelt==1.0.0
rich==14.1.0
rich-toolkit==0.14.9
rignore==0.6.4
rsa==4.9.1
sentry-sdk==2.34.0
setuptools==69.1.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sorl-thumbnail==12.10.0
soupsieve==2.5
sqladmin==0.21.0
SQLAlchemy==2.0.42
sqlparse==0.4.4
starlette==0.47.2
svglib==1.5.1
tasks==2.8.0
tinycss2==1.2.1
tomlkit==0.12.4
trove-classifiers==2024.3.25
typer==0.16.0
types-beautifulsoup4==4.12.0.20240229
types-html5lib==1.1.11.20240228
types-PyYAML==6.0.12.12
types-requests==2.31.0.20240310
typing-inspection==0.4.1
typing_extensions==4.14.1
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
virtualenv==20.25.1
watchfiles==1.1.0
webencodings==0.5.1
websockets==15.0.1
WTForms==3.1.2
xattr==1.1.0
